Inlined OverviewPresenterImpl
