+++ src/main/java/de/vogel612/helper/Main.java
@@ -1,9 +1,8 @@
 package de.vogel612.helper;
 
-import de.vogel612.helper.ui.OverviewPresenter;
 import de.vogel612.helper.ui.OverviewView;
 import de.vogel612.helper.ui.OverviewModel;
-import de.vogel612.helper.ui.OverviewPresenterImpl;
+import de.vogel612.helper.ui.OverviewPresenter;
 import de.vogel612.helper.ui.OverviewViewImpl;
 import de.vogel612.helper.ui.TranslationPresenter;
 
@@ -39,7 +38,7 @@ public class Main {
         OverviewModel m = new OverviewModel();
         OverviewView v = new OverviewViewImpl();
 
-        OverviewPresenter p = new OverviewPresenterImpl(m, v, tp);
+        OverviewPresenter p = new OverviewPresenter(m, v, tp);
         p.initialize();
         p.loadFiles(resxFolder);
         p.show();
